Without serialization at the end of block function, the blocks match test vectors
Test vectors: https://tools.ietf.org/html/rfc7539

For some reason, in the algorithm description it is said that after
addition of original state, the state should be serialized (each entry written as
little endian) but when I do this, the hashed blocks don't match the ones in the test vectors.

When I don't serialize, they do match. What is going on here?