Fix version settings selection not supporting custom type
I wanted to set a variable of type ChaCha.Version with a settings combobox.

But apparently, only the index of the selected element is returned and casted to a ChaCha.Version (somehow and if I understood correctly what was happening). Thus, I can not directly set a ChaCha.Version variable.

Instead, I created a "wrapper variable" which is of type int and is set during settings selection. I used a switch to then set the actual version variable of type ChaCha.Version to the correct version depending on the given selected index.