Mark ActionNavigation API for automatic undoing as obsolete
Unfortunately, the API for automatic undoing is not compatible with caching. This is so because to use the automatic undoing, the page action needs to be executed to generate the undo action and push it onto the undo stack.
It works by saving the state of the UI elements which have been changed during page execution. But since caching skips all these page actions to move in constant time to a cached action index this no longer works and must be replaced with manual undo actions :(