Remove ExtendLastActionWithRoundSetter
Removed because approach was not good and error-prone.

I should add a setter action to the sequence without adding it to the
page actions to prevent the setter being a own page action.

This way, all next actions in the sequence will be extended
automatically with that setter actions and added extended to the page
actions.