Revert workaround attempt of last 3 commits
I tried to reinitialize the user input fields in the ChaCha Hash view on new workspace execution WHILE we are on that page.

But this turned out to be more hackier than I thought.

I should just force the user back to the first page. Then everything
should work smoothly.