Prepare architecture change by value renaming
I will essentially need two wrapper variables like I did for the action
input textbox binding.

Both the diffusion input and the diffusion xor input have a two-way binding to it's wrapper variable to
enable validation.

If the focus is lost or the enter key is pressed, a function is executed
which updates the underlying variable. This then notifies the wrapper
variables of the change thus reflecting the change in the UI via the
two-way binding.